add basic user registration and validation schemas for posts and user

+++ controller/postContoller.js
@@ -1,7 +1,14 @@
-const Post = require( "../models/Post" )
-const catchAsync = require("../utils/catchAsync")
+const mongoose = require( "mongoose" );
+const Post = require( "../models/Post" );
+const { postInsertSchema, postUpdateSchema } = require( "../schemas/postSchema" );
+const catchAsync = require("../utils/catchAsync");
+const CustomError = require( "../utils/CustomError" );
 
 const insertPost = catchAsync(async(req,res,next)=>{
+    const { error } = postInsertSchema.validate(req.body);
+  if (error) {
+    return next(new CustomError(400, error.details[0].message));
+  }
     const {title,content} = req.body;
     const post = Post({title,content});
     await post.save()
@@ -13,14 +20,26 @@ const getAllPosts = catchAsync(async(req,res,next)=>{
     res.status(201).json({status:"success",data:posts})
 })
 
+const getPost = catchAsync(async(req,res,next)=>{
+    if(!mongoose.Types.ObjectId.isValid(req.params.id)){return next(new CustomError(400,"Invalid Id"))}
+    const posts =await Post.findById(req.params.id).select(["-__v"])
+    res.status(201).json({status:"success",data:posts})
+})
+
 const updatePost = catchAsync(async(req,res,next)=>{
+    const { error } = postUpdateSchema.validate(req.body);
+  if (error) {
+    return next(new CustomError(400, error.details[0].message));
+  }
+    if(!mongoose.Types.ObjectId.isValid(req.params.id)){return next(new CustomError(400,"Invalid Id"))}
     const post =await Post.updateOne({_id:req.params.id},req.body)
     res.status(201).json({status:"success",data:post})
 })
 
 const deletePost = catchAsync(async(req,res,next)=>{
+    if(!mongoose.Types.ObjectId.isValid(req.params.id)){return next(new CustomError(400,"Invalid Id"))}
     const post =await Post.deleteOne({_id:req.params.id})
     res.status(201).json({status:"success",data:post})
 })
 
-module.exports = {insertPost,getAllPosts,deletePost,updatePost}
+module.exports = {insertPost,getAllPosts,deletePost,updatePost,getPost}
